backend/requirements.txt
fastapi
uvicorn
motor
bcrypt==3.2.2
python-jose[cryptography]
passlib[bcrypt]
pydantic[email]
python-dotenv
